requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.43.0
django-extra-views==0.13.0
django-haystack==3.0
django-jalali==4.0.0
django-oscar==2.1
django-phonenumber-field==3.0.1
django-tables2==2.2.1
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
environ==1.0
example==0.1.0
factory-boy==2.12.0
Faker==4.14.0
gunicorn==20.0.4
idna==2.10
jdatetime==3.6.2
mysqlclient==2.0.1
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.2
phonenumbers==8.12.11
Pillow==8.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
purl==1.5
pyAHP==0.1.2
pyanp==0.3.1
PyAudio @ file:///C:/Users/Administrator/PycharmProjects/General/shop/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
pycountry==20.7.3
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
scipy==1.5.2
six==1.15.0
sorl-thumbnail==12.6.3
SpeechRecognition==3.8.1
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
whitenoise==5.2.0
